refactor(step1): index, app 이외의 모든 파일 위치 이동
Checker -> domain
utils.js 중 readLineAsync 함수 -> UI
utils.js 삭제
view를 step1로 분리
