+++ src/domain/Checkers.js
