+++ src/view/step1/InputView.js
@@ -1,4 +1,4 @@
-import { readLineAsync } from "../utils.js";
+import readLineAsync from "./readLineAsync.js";
 
 class InputView {
   static async getInput(message) {
