+++ src/view/step1/OutputView.js
